[refactor] Use camelCase for fungibleAsset
