+++ docs/@aptos-labs/ts-sdk-0.0.0/classes/AptosConfig.html
@@ -23,7 +23,7 @@
 <ul class="tsd-hierarchy">
 <li><span class="target">AptosConfig</span></li></ul></section><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L11">src/api/aptos_config.ts:11</a></li></ul></aside>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L11">src/api/aptosConfig.ts:11</a></li></ul></aside>
 <section class="tsd-panel-group tsd-index-group">
 <section class="tsd-panel tsd-index-panel">
 <details class="tsd-index-content tsd-index-accordion" open><summary class="tsd-accordion-summary tsd-index-summary">
@@ -55,42 +55,42 @@
 <h5><code class="tsd-tag ts-flagOptional">Optional</code> settings: <a href="../types/AptosSettings.html" class="tsd-signature-type" data-tsd-kind="Type alias">AptosSettings</a></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <a href="AptosConfig.html" class="tsd-signature-type" data-tsd-kind="Class">AptosConfig</a></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L32">src/api/aptos_config.ts:32</a></li></ul></aside></li></ul></section></section>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L32">src/api/aptosConfig.ts:32</a></li></ul></aside></li></ul></section></section>
 <section class="tsd-panel-group tsd-member-group">
 <h2>Properties</h2>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="clientConfig" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <code class="tsd-tag ts-flagReadonly">Readonly</code> <span>client<wbr/>Config</span><a href="#clientConfig" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">client<wbr/>Config<span class="tsd-signature-symbol">?:</span> <a href="../types/ClientConfig.html" class="tsd-signature-type" data-tsd-kind="Type alias">ClientConfig</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L30">src/api/aptos_config.ts:30</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L30">src/api/aptosConfig.ts:30</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="faucet" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <code class="tsd-tag ts-flagReadonly">Readonly</code> <span>faucet</span><a href="#faucet" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">faucet<span class="tsd-signature-symbol">?:</span> <span class="tsd-signature-type">string</span></div>
 <div class="tsd-comment tsd-typography"><p>The optional hardcoded faucet URL to send requests to instead of using the network</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L23">src/api/aptos_config.ts:23</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L23">src/api/aptosConfig.ts:23</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="fullnode" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <code class="tsd-tag ts-flagReadonly">Readonly</code> <span>fullnode</span><a href="#fullnode" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">fullnode<span class="tsd-signature-symbol">?:</span> <span class="tsd-signature-type">string</span></div>
 <div class="tsd-comment tsd-typography"><p>The optional hardcoded fullnode URL to send requests to instead of using the network</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L18">src/api/aptos_config.ts:18</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L18">src/api/aptosConfig.ts:18</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="indexer" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <code class="tsd-tag ts-flagReadonly">Readonly</code> <span>indexer</span><a href="#indexer" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">indexer<span class="tsd-signature-symbol">?:</span> <span class="tsd-signature-type">string</span></div>
 <div class="tsd-comment tsd-typography"><p>The optional hardcoded indexer URL to send requests to instead of using the network</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L28">src/api/aptos_config.ts:28</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L28">src/api/aptosConfig.ts:28</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="network" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>network</span><a href="#network" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">network<span class="tsd-signature-symbol">:</span> <a href="../enums/Network.html" class="tsd-signature-type" data-tsd-kind="Enumeration">Network</a></div>
 <div class="tsd-comment tsd-typography"><p>The Network that this SDK is associated with.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L13">src/api/aptos_config.ts:13</a></li></ul></aside></section></section></div>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/api/aptos_config.ts#L13">src/api/aptosConfig.ts:13</a></li></ul></aside></section></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
